What i learned: - now i'm more confident with using a weak self referencing
briefly
it is useful the our app request data that will be pass in a view.
Then we create an instance of the class ARC system counts how many references app created.
lets imagine that our code request data from the internet with delay. If we will use strong reference (a.k.a. self.) and decide to dismiss the object that requested this data. ARC will still remember that there is an instance of our class and will not deinitilaize it until it will not receive that data.

To avoid it, use a weak reference to this class to be able cancel asynchronous task and deinit the instance freely. It's more optimise way to use memory and avoid creating uncountable amount of instances that can load CPU and Memory